fix: separate computer playing class
