+++ chess/chess.py
@@ -1,6 +1,7 @@
 from random import choice
 
 from piece import Bishop, King, Knight, Pawn, Queen, Rook
+from hai import Hai
 
 WHITE = "white"
 BLACK = "black"
@@ -11,7 +12,7 @@ uniDict = {
 }
 
 
-class Game:
+class Game(Hai):
     def __init__(self):
         self.playersturn = BLACK
         self.message = "this is where prompts will go"
@@ -144,32 +145,6 @@ class Game:
         except:
             print("error decoding input. please try again")
             return ((-1, -1), (-1, -1))
-
-    def autoInput(self, Color):
-        piecelist = []
-        for position, piece in self.gameboard.items():
-            if piece.Color == self.playersturn:
-                piecelist.append((piece, position))
-        for _ in range(100):
-            rand_start = choice(piecelist)
-            position = rand_start[1]
-            piece = rand_start[0]
-            available_list = piece.availableMoves(
-                position[0], position[1], self.gameboard, Color
-            )
-            if not available_list:
-                continue
-            rand_end = choice(available_list)
-
-            print("computer played", position, rand_end)
-
-            return position, rand_end
-
-    """def validateInput(self, *kargs):
-        for arg in kargs:
-            if type(arg[0]) is not type(1) or type(arg[1]) is not type(1):
-                return False
-        return True"""
 
     def printBoard(self):
         print("   1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 |")
